+++ EssentialFeediOS/Feed UI/Composers/FeedUIComposer.swift
@@ -40,24 +40,3 @@ private extension FeedUIComposer {
         return feedController
     }
 }
-
-private final class FeedViewAdapter: FeedView {
-    private weak var controller: FeedViewController?
-    private let loader: FeedImageDataLoader
-
-    init(controller: FeedViewController, loader: FeedImageDataLoader) {
-        self.controller = controller
-        self.loader = loader
-    }
-
-    func display(_ viewModel: FeedViewModel) {
-        controller?.tableModel = viewModel.feed.map { model in
-            let adapter = FeedImageDataLoaderPresentationAdapter<WeakRefVirualProxy<FeedImageCellController>, UIImage>(model: model, loader: loader)
-
-            let view = FeedImageCellController(delegate: adapter)
-            adapter.presenter = FeedImagePresenter(view: WeakRefVirualProxy(view), imageTransformer: UIImage.init)
-
-            return view
-        }
-    }
-}
